feat(181): move ProcessRunner and CommandNormalizationStrategy into JD.Efcpt.Build.Core
ProcessRunner/ProcessResult are now public (JD.Efcpt.Build.Core.Processes)
since RunEfcpt, EnsureDacpacBuilt, and DefaultToolAcquirer call them
across the assembly boundary; CommandNormalizationStrategy/ProcessCommand
stay internal (only used internally by ProcessRunner), covered by the
InternalsVisibleTo added in the previous commit for test access.

All 4 TFMs of Tasks build clean; JD.Efcpt.Build.Tests builds clean and
targeted RunEfcpt/EnsureDacpacBuilt/OfflineMode/EfcptDoctor tests (84)
pass unchanged.
